[#1187] Split Onetime module into specialized helpers
* Splits monolithic Onetime module into logical helper modules
* Moves related functionality into environment, database, locales,
  mail, globals, and initialization helpers
